# docs/.gitbook/assets/openapi.yaml
openapi: 3.0.0
paths:
  /cats:
    post:
      operationId: create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCatDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
      tags:
        &a1
        - cats
      security:
        &a2
        - bearer: []
  "/cats/{id}":
    get:
      operationId: findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cat"
      tags: *a1
      security: *a2
  /cats/test:
    get:
      operationId: testEndpoint
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: string
      tags: *a1
      security: *a2
  /cats/add:
    get:
      operationId: testApiAppend
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: string
      tags: *a1
      security: *a2
info:
  title: Cats example
  description: The cats API description
  version: "1.0"
  contact: {}
tags:
  - name: cats
    description: ""
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateCatDto:
      type: object
      properties:
        name:
          type: string
        age:
          type: number
        breed:
          type: string
      required:
        - name
        - age
        - breed
    Cat:
      type: object
      properties:
        name:
          type: string
        age:
          type: number
        breed:
          type: string
      required:
        - name
        - age
        - breed